+++ src/torchmetrics/functional/text/__init__.py
@@ -24,10 +24,11 @@ from torchmetrics.functional.text.ter import translation_edit_rate  # noqa: F401
 from torchmetrics.functional.text.wer import word_error_rate  # noqa: F401
 from torchmetrics.functional.text.wil import word_information_lost  # noqa: F401
 from torchmetrics.functional.text.wip import word_information_preserved  # noqa: F401
-from torchmetrics.utilities.imports import _NLTK_AVAILABLE, _TRANSFORMERS_AUTO_AVAILABLE
+from torchmetrics.utilities.imports import _NLTK_AVAILABLE, _TRANSFORMERS_AVAILABLE
 
-if _TRANSFORMERS_AUTO_AVAILABLE:
+if _TRANSFORMERS_AVAILABLE:
     from torchmetrics.functional.text.bert import bert_score  # noqa: F401
+    from torchmetrics.functional.text.infolm import infolm  # noqa: F401
 
 if _NLTK_AVAILABLE:
     from torchmetrics.functional.text.rouge import rouge_score  # noqa: F401